add hotkey dialog function to accept the first available quest (or reward)
note: this currently only works if the dialog window is already open, implement a queue later? where, HotkeyDialog or in DialogModule?

+++ GWToolboxdll/Windows/Hotkeys.h
@@ -331,7 +331,7 @@ public:
 
 class HotkeyDialog : public TBHotkey {
 public:
-    size_t id = 0;
+    uint32_t id = 0;
     char name[140];
 
     static const char* IniSection() { return "Dialog"; }
